consolidate common number format functions and market icon pair component
