+++ src/typescript/frontend/src/components/trade/DepositWithdrawModal/SelectMarketContent.tsx
@@ -7,13 +7,13 @@ import {
   getFilteredRowModel,
   useReactTable,
 } from "@tanstack/react-table";
-import Image from "next/image";
-import { useEffect, useMemo, useState } from "react";
+import { useMemo, useState } from "react";
 
+import { MarketIconPair } from "@/components/MarketIconPair";
 import { type ApiMarket, type ApiStats } from "@/types/api";
+import { formatNumber, plusMinus } from "@/utils/formatter";
 
 import { useAllMarketData, useAllMarketPrices, useAllMarketStats } from ".";
-
 const columnHelper = createColumnHelper<ApiMarket>();
 
 const TABLE_SPACING = {
@@ -364,85 +364,4 @@ const getPriceByMarketId = (
 ) => {
   if (!prices) return undefined;
   return prices.find((price) => price.market_id === marketId);
-};
-
-// copy paste from statsbar, think about making a unified component later
-const DEFAULT_TOKEN_ICON = "/tokenImages/default.png";
-type MarketIconPairProps = {
-  baseAssetIcon?: string;
-  quoteAssetIcon?: string;
-};
-const MarketIconPair = ({
-  baseAssetIcon = DEFAULT_TOKEN_ICON,
-  quoteAssetIcon = DEFAULT_TOKEN_ICON,
-}: MarketIconPairProps) => {
-  // TODO: add this so statsbar can use it too
-  interface ImageWithFallbackProps {
-    fallback?: string;
-    alt: string;
-    src: string;
-    [key: string]: any; // allow any other props
-  }
-
-  const ImageWithFallback: React.FC<ImageWithFallbackProps> = ({
-    fallback = "/tokenImages/default.png",
-    alt,
-    src,
-    ...props
-  }) => {
-    const [error, setError] = useState<Error | null>(null);
-
-    useEffect(() => {
-      setError(null);
-    }, [src]);
-
-    return (
-      <Image
-        alt={alt}
-        onError={() => setError(new Error("Failed to load image"))}
-        src={error ? fallback : src}
-        {...props}
-      />
-    );
-  };
-
-  return (
-    <div className="relative flex">
-      {/* height width props required */}
-      <ImageWithFallback
-        src={baseAssetIcon}
-        alt="market-icon-pair"
-        width={40}
-        height={40}
-        className="z-20 aspect-square  w-[30px] min-w-[30px] md:min-w-[40px]"
-      ></ImageWithFallback>
-      <ImageWithFallback
-        src={quoteAssetIcon}
-        alt="market-icon-pair"
-        width={40}
-        height={40}
-        className="absolute z-10 aspect-square w-[30px] min-w-[30px] translate-x-1/2 md:min-w-[40px]"
-      ></ImageWithFallback>
-    </div>
-  );
-};
-
-const getLang = () => {
-  return typeof window === "undefined"
-    ? "en"
-    : navigator.language || navigator.languages[0];
-};
-
-const formatNumber = (num: number | undefined, digits: number): string => {
-  if (!num) return "-";
-  return num.toLocaleString(getLang(), {
-    minimumFractionDigits: digits,
-    maximumFractionDigits: digits,
-  });
-};
-
-const plusMinus = (num: number | undefined): string => {
-  if (!num) return "";
-  // no need to return - as numbers will already have that
-  return num >= 0 ? `+` : ``;
 };
